Updated for most tests to run on Windows
